+++ src/components/styles/StyledButton.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+import startButton from '@assets/startButton.svg';
+
+const StyledButton = styled.button`
+  background-image: url("${startButton}");
+  margin: 25px;
+`;
+
+export default StyledButton;